backend/requirements.txt
curl-cffi==0.7.4
fastapi==0.115.5
uvicorn[standard]==0.32.0
yfinance==0.2.50
feedparser==6.0.11
praw==7.8.1
httpx==0.28.0
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic==2.10.3
anthropic==0.40.0
